refactor acp - splitted, refactor core loop
